* commit 'b27e99aa038638a1dcea8316261f5059bf98375d':
  public/src/core/ISessionWatchdog.h edited online with Bitbucket
  public/src/core/ISessionWatchdog.h edited online with Bitbucket
  public/src/core/util/IInterruptibleThreadSuspender.h edited online with Bitbucket
  Add watchdog for not immediatly closable sessions

GitOrigin-RevId: 268f9e3a2b54136c39d9d7a85a48dd7919f3730f

+++ src/api-c/CustomLogger.cxx
@@ -111,7 +111,7 @@ void CustomLogger::doLog(LOG_LEVEL level, const char* format, va_list args)
 	// perform the printf argument filling
 	va_list argcopy;
 	va_copy(argcopy, args);
-	int length = vsnprintf(nullptr, 0, format, argcopy);	// excl. term. 0
+	size_t length = vsnprintf(nullptr, 0, format, argcopy);	// excl. term. 0
 	va_end(argcopy);
 	char *traceStatement = new char[length + 1]; // + 1 for term. 0
 	vsnprintf(traceStatement, length + 1, format, args);
